requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
Flask==2.1.2
idna==3.3
importlib-metadata==4.11.3
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.1
ndg-httpsclient==0.5.1
numpy==1.22.3
pandas==1.4.2
py4j==0.10.9.3
pyasn1==0.4.8
pycparser==2.21
pymongo==4.1.1
pyOpenSSL==22.0.0
pyspark==3.2.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
